// src/app/model/superpave-gradation.ts
export interface SuperpaveGradation {
  id: number;
  standard: string;

  cumRetainedA: number;
  cumRetainedB: number;
  cumRetainedC: number;
  cumRetainedD: number;
  cumRetainedE: number;
  cumRetainedF: number;
  cumRetainedG: number;
  cumRetainedH: number;
  cumRetainedI: number;
  cumRetainedJ: number;
  cumRetainedK: number;
  cumRetainedL: number;
  cumRetainedM: number;

  jmfA: string;
  jmfB: string;
  jmfC: string;
  jmfD: string;
  jmfE: string;
  jmfF: string;
  jmfG: string;
  jmfH: string;
  jmfI: string;
  jmfJ: string;
  jmfK: string;
  jmfL: string;
  jmfM: string;

  controlPointA: string;
  controlPointB: string;
  controlPointC: string;
  controlPointD: string;
  controlPointE: string;
  controlPointF: string;
  controlPointG: string;
  controlPointH: string;
  controlPointI: string;
  controlPointJ: string;
  controlPointK: string;
  controlPointL: string;
  controlPointM: string;

  expandA: string;
  expandB: string;
  expandC: string;
  expandD: string;
  expandE: string;
  expandF: string;
  expandG: string;
  expandH: string;
  expandI: string;
  expandJ: string;
  expandK: string;
  expandL: string;
  expandM: string;

  [key: string]: any;
}